# LegalConnect.API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "LegalConnect API",
    "description": "Lawyer Marketplace Platform API",
    "version": "v1"
  },
  "paths": {
...
    "/api/admin/lawyers/{lawyerId}/approve": {
      "put": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "lawyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/lawyers/{lawyerId}/reject": {
      "put": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "lawyerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RejectLawyerDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/commission": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SetCommissionDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/categories": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCategoryDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/categories/{id}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCategoryDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/appointments/slots": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lawyerId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "date",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/appointments": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BookAppointmentDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/appointments/my": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/appointments/{id}": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/appointments/{id}/cancel": {
      "put": {
        "tags": [
          "Appointments"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CancelAppointmentDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/appointments/lawyer": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "type": "string"
          },
          {
            "in": "query",
            "name": "to",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/appointments/{id}/confirm": {
      "put": {
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/appointments/{id}/complete": {
      "put": {
        "tags": [
          "Appointments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/lawyers": {
      "get": {
        "tags": [
          "Lawyers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "SearchTerm",
            "type": "string"
          },
          {
            "in": "query",
            "name": "City",
            "type": "string"
          },
          {
            "in": "query",
            "name": "CategoryId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "Court",
            "type": "string"
          },
          {
            "in": "query",
            "name": "MinExperience",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "MaxExperience",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "MinFee",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "MaxFee",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "MinRating",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "IsVerified",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "IsAvailable",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "SortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SortDescending",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BookAppointmentDto": {
      "required": [
        "appointmentDate",
        "lawyerId",
        "startTime"
      ],
      "type": "object",
      "properties": {
        "lawyerId": {
          "format": "int32",
          "type": "integer"
        },
        "appointmentDate": {
          "format": "date-time",
          "type": "string"
        },
        "startTime": {
          "format": "date-span",
          "type": "string"
        },
        "notes": {
          "maxLength": 500,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CancelAppointmentDto": {
      "required": [
        "reason"
      ],
      "type": "object",
      "properties": {
        "reason": {
          "maxLength": 300,
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "CreateCategoryDto": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "description": {
          "maxLength": 255,
          "type": "string"
        },
        "iconClass": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "RejectLawyerDto": {
      "required": [
        "reason"
      ],
      "type": "object",
      "properties": {
        "reason": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "SetCommissionDto": {
      "type": "object",
      "properties": {
        "commissionPercentage": {
          "format": "double",
          "maximum": 100,
          "minimum": 0,
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateCategoryDto": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "description": {
          "maxLength": 255,
          "type": "string"
        },
        "iconClass": {
          "type": "string"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}